+++ tests/TestCase.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ntoufoudis\Hopper\Tests;
+
+use Illuminate\Foundation\Application;
+use Ntoufoudis\Hopper\HopperServiceProvider;
+use Orchestra\Testbench\TestCase as Orchestra;
+
+abstract class TestCase extends Orchestra
+{
+    /**
+     * @param  Application  $app
+     * @return array<int, class-string>
+     */
+    protected function getPackageProviders($app): array
+    {
+        return [
+            HopperServiceProvider::class,
+        ];
+    }
+}